more tests - unclear whether these stubs are actually used?
