+++ package.json
@@ -44,10 +44,10 @@
     "sonar-scan": "node_modules/sonar-scanner/bin/sonar-scanner -Dsonar.projectName='probate-caveats-frontend' -Dsonar.sources=app -Dsonar.cpd.exclusions=app/utils/FeatureToggle.js,app/utils/FormatName.js,app/core/runners/ActionStepRunner.js,app/**/template.html,app/**/scripts.html -Dsonar.exclusions=app/assets/**,app/**/scripts.html,app/steps/ui/static/cookies/index.js,app/steps/ui/equality/createToken.js -Dsonar.language=js -Dsonar.javascript.lcov.reportPaths=coverage/lcov.info",
     "git-info": "node-git-info-json",
     "//": "The below scripts are just stubs, don't use them in production",
-    "stub:orchestrator": "NODE_PATH=. node test/service-stubs/orchestrator.js",
-    "stub:s2s": "NODE_PATH=. node test/service-stubs/s2s.js",
-    "stub:idam": "NODE_PATH=. node test/service-stubs/idam.js",
-    "stub:equality-and-diversity-endpoint": "NODE_PATH=. node test/service-stubs/equalityAndDiversityEndpoint.js",
+    "stub:orchestrator": "NODE_PATH=. node test/service-stubs/orchestrator.cjs",
+    "stub:s2s": "NODE_PATH=. node test/service-stubs/s2s.cjs",
+    "stub:idam": "NODE_PATH=. node test/service-stubs/idam.cjs",
+    "stub:equality-and-diversity-endpoint": "NODE_PATH=. node test/service-stubs/equalityAndDiversityEndpoint.cjs",
     "fortifyScan": "./test/java/gradlew -p test/java fortifyScan",
     "stub:equality-and-diversity-health": "NODE_PATH=. node test/service-stubs/equalityAndDiversityHealth.js",
     "suppress-cve": "yarn npm audit --recursive --environment production --json > yarn-audit-known-issues"
