+++ test/service-stubs/s2s.cjs
@@ -1,22 +1,20 @@
 'use strict';
 
 /* eslint no-console: 0 */
+/* eslint import/no-commonjs: 0 */
 // idam service-service auth stub
 
 const express = require('express');
 const app = express();
 const router = express.Router();
-const IDAM_STUB_PORT = 4501;
+const S2S_STUB_PORT = 4502;
 
-router.post('/oauth2/authorize', (req, res) => {
+router.post('/lease', (req, res) => {
+    console.log(req.headers);
+    console.log(req.body);
+
     res.status(200);
-    res.send({code: 'codeValue'});
-
-});
-
-router.post('/oauth2/token', (req, res) => {
-    res.status(200);
-    res.send({access_token: 'accessToken'});
+    res.send('eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJSRUZFUkVOQ0UifQ.Z_YYn0go02ApdSMfbehsLXXbxJxLugPG8v_3ktCpQurK8tHkOy1qGyTo02bTdilX4fq4M5glFh80edDuhDJXPA');
 
 });
 
@@ -29,7 +27,7 @@ router.get('/health', (req, res) => {
 app.use(router);
 
 // start the app
-console.log(`Listening on: ${IDAM_STUB_PORT}`);
-const server = app.listen(IDAM_STUB_PORT);
+console.log(`Listening on: ${S2S_STUB_PORT}`);
+const server = app.listen(S2S_STUB_PORT);
 
 module.exports = server;
