+++ test/step_file.js
@@ -1,7 +1,6 @@
-'use strict';
 // in this file you can append custom step methods to 'I' object
 
-module.exports = () => {
+export default () => {
     return actor({
 
         // Define custom steps here, use 'this' to access default methods of I.
